refactor: move task interface into task module
