+++ src/app/add-task/add-task.component.ts
@@ -7,7 +7,7 @@ import {
 import { Subscription } from 'rxjs';
 
 import { UiService } from 'src/app/shared/services/ui.service';
-import { Task } from 'src/app/task';
+import { Task } from '../task/task.model';
 
 @Component({
   selector: 'app-add-task',
