+++ src/app/tasks/tasks.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
-import { Task } from '../task';
+import { Task } from '../task/task.model';
 
 const headers = new HttpHeaders({
   'content-type': 'application/json',
